fastapi-sqlachemy-docker/requirements.txt
alembic==1.5.5
click==7.1.2
fastapi==0.63.0
FastAPI-SQLAlchemy==0.2.1
h11==0.12.0
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
pydantic==1.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
uvicorn==0.13.4
